Credits fjernes ved brug af prompt, man kan ikke køre prompt med 0 credits

+++ src/main/java/dat3/openai_demo/api/MealMentorController.java
@@ -23,27 +23,43 @@ public class MealMentorController {
   }
 
 
-  @PostMapping()
+
+    @PostMapping()
       public MyResponse generatePrompt(@RequestBody UserPromptResponse userPromptResponse){
 
-      String usernamePrompting = userPromptResponse.getUsername();
-      UserResponse user = userService.getUser(usernamePrompting);
+            // Check if the user has credits
+        if (userService.getUser(userPromptResponse.getUsername()).getCredits() <= 0) {
 
-      userService.removeOneCreditFromUser(usernamePrompting);
+            return new MyResponse("{\n" +
+                    "    \"Message\": \"USER HAS NO CREDITS\"\n" +
+                    "}");
+
+        } else {
+
+            // Removing one credit from the user's credits
+            String usernamePrompting = userPromptResponse.getUsername();
+            UserResponse user = userService.getUser(usernamePrompting);
+            userService.removeOneCreditFromUser(usernamePrompting);
+
+            // Printing how many cr
+            System.out.println("****** " + user.getUsername() + " is PROMPTING *******\n" +
+                    "****** and he/she has got " +
+                    userService.getUser(usernamePrompting).getCredits()
+                    + " credits left *****");
+
+            // The userPrompt
+            String userPrompt = "I am a " + userPromptResponse.getUserAge() + " old "
+                    + userPromptResponse.getSex() +
+                    "and " + "i workout" + userPromptResponse.getWorkoutsPerWeek() + "times a week " +
+                    "The recipes should include" + userPromptResponse.getMealChecklist() + "and must not include " + userPromptResponse.getPreferences()
+                    + ". My goals are" + userPromptResponse.getGoals()
+
+                    + " i want the recipe made for " /* + userPromptResponse.getAmountOfDays() + "number of days" */ ;
+
+            return openAiService.makeRequest(userPrompt, SYSTEM_MESSAGE);
+        }
 
 
-      System.out.println("****** " + user.getUsername() + " is PROMPTING *******\n" +
-              "****** and he/she has got " + userService.getUser(usernamePrompting).getCredits() + " credits left *****");
-
-      String userPrompt = "I am a " + userPromptResponse.getUserAge() + " old "
-              + userPromptResponse.getSex() +
-              "and " + "i workout" + userPromptResponse.getWorkoutsPerWeek() + "times a week " +
-              "The recipes should include" + userPromptResponse.getMealChecklist() + "and must not include " + userPromptResponse.getPreferences()
-              + ". My goals are" + userPromptResponse.getGoals()
-
-               + " i want the recipe made for " /* + userPromptResponse.getAmountOfDays() + "number of days" */ ;
-
-      return openAiService.makeRequest(userPrompt, SYSTEM_MESSAGE);
 
   }
